[TWINFACES-412] extracted createdAt into two createdAtFrom and createdAtTo
