+++ src/app/core/twins/[twinId]/page.tsx
@@ -1,5 +1,16 @@
+import { TwinContextProvider } from "@/features/twin";
 import { TwinScreen } from "@/screens/twin";
 
-export default function Page() {
-  return <TwinScreen />;
+type Props = {
+  params: {
+    twinId: string;
+  };
+};
+
+export default function Page({ params: { twinId } }: Props) {
+  return (
+    <TwinContextProvider twinId={twinId}>
+      <TwinScreen />
+    </TwinContextProvider>
+  );
 }
