--- src/components/form-fields/calendar/index.ts
@@ -1,1 +1,0 @@
-export * from "./calendar-form-field";
