--- src/entities/attachment/api/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from "./use-search-v1";
-export * from "./use-fetch-by-id";
