--- src/entities/attachment/api/index.ts
@@ -1,3 +1,0 @@
-export * from "./hooks";
-export * from "./api-service";
-export * from "./types";
