--- src/entities/attachment/index.ts
@@ -1,2 +1,0 @@
-export * from "./api";
-export * from "./libs";
