+++ src/entities/twin/libs/hooks/index.ts
@@ -1,2 +1,2 @@
 export * from "./use-select-adapter";
-export * from "./useFilters";
+export * from "./use-filters";
