+++ src/entities/twin/server/api/types.ts
@@ -1,4 +1,8 @@
+import { Comment_DETAILED } from "@/entities/comment";
+import { Permission } from "@/entities/permission";
 import { TwinClass_DETAILED } from "@/entities/twin-class";
+import { TwinClassField_DETAILED } from "@/entities/twin-class-field";
+import { TwinFlowTransition_DETAILED } from "@/entities/twin-flow-transition";
 import { User } from "@/entities/user";
 import { components, operations } from "@/shared/api/generated/schema";
 import { RequireFields } from "@/shared/libs";
@@ -50,3 +54,47 @@ export type TwinFilters = Partial<
 >;
 
 export type TwinSimpleFilters = components["schemas"]["TwinSearchSimpleV1"];
+
+export type Attachment = components["schemas"]["AttachmentV1"];
+export type Attachment_DETAILED = Required<
+  Attachment & {
+    twin: Twin;
+    twinClassField: TwinClassField_DETAILED;
+    twinflowTransition: TwinFlowTransition_DETAILED;
+    viewPermission: Permission;
+    authorUser: User;
+    comment: Comment_DETAILED;
+  }
+>;
+
+export type AttachmentSearchRqV1 =
+  components["schemas"]["AttachmentSearchRqV1"];
+
+export type AttachmentRqQuery =
+  operations["attachmentViewV1"]["parameters"]["query"];
+
+export type AttachmentFilterKeys =
+  | "idList"
+  | "twinIdList"
+  | "externalIdLikeList"
+  | "twinflowTransitionIdList"
+  | "storageLinkLikeList"
+  | "createdByUserIdList"
+  | "titleLikeList"
+  | "descriptionLikeList"
+  | "viewPermissionIdList"
+  | "twinClassFieldIdList"
+  | "createdAtFrom"
+  | "createdAtTo";
+
+export type AttachmentFilters = Partial<
+  Pick<
+    components["schemas"]["AttachmentSearchRqV1"] & {
+      createdAtFrom: string;
+      createdAtTo: string;
+    },
+    AttachmentFilterKeys
+  >
+>;
+
+export type DataTimeRangeV1 = components["schemas"]["DataTimeRangeV1"];
