+++ src/features/api/context/private-context-provider.tsx
@@ -6,7 +6,6 @@ import {
   PermissionAssigneePropagationApi,
   createPermissionAssigneePropagationApi,
 } from "@/entities/assigneePropagation";
-import { AttachmentApi, createAttachmentApi } from "@/entities/attachment";
 import { CommentApi, createCommentApi } from "@/entities/comment";
 import { DatalistApi, createDatalistApi } from "@/entities/datalist";
 import {
@@ -117,7 +116,6 @@ export interface PrivateApiContextProps {
   datalistOption: DatalistOptionApi;
   link: LinkApi;
   tier: TierApi;
-  attachment: AttachmentApi;
 }
 
 export function PrivateApiContextProvider({
@@ -166,7 +164,6 @@ export function PrivateApiContextProvider({
         datalistOption: createDatalistOptionApi(settings),
         link: createLinkApi(settings),
         tier: createTierApi(settings),
-        attachment: createAttachmentApi(settings),
       }
     : ({} as PrivateApiContextProps);
