+++ src/screens/attachment/index.ts
@@ -1,0 +1,2 @@
+export * from "./attachment";
+export * from "./views";
